backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.2",
    "google-auth-httplib2>=0.3.1",
    "gspread>=6.2.1",
    "openai>=2.32.0",
    "requests>=2.33.1",
]
